sistana: enable dataclass slots=True or hand-written
