+++ src/arclet/alconna/sistana/model/track.py
@@ -4,16 +4,18 @@ from collections import deque
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
+from arclet.alconna._dcls import safe_dcls_kw
+
 from ..err import ReceivePanic, TransformPanic, ValidateRejected
 from ..utils.misc import Value
 from .fragment import _Fragment, assert_fragments_order
 
 if TYPE_CHECKING:
-    from ..buffer import Buffer
+    from elaina_segment import Buffer
     from .receiver import RxGet, RxPut
 
 
-@dataclass
+@dataclass(**safe_dcls_kw(slots=True))
 class Track:
     fragments: deque[_Fragment]
     assignes: dict[str, Any] = field(default_factory=dict)
@@ -41,6 +43,32 @@ class Track:
             first = self.fragments[-1]
             if first.variadic and first.name not in self.assignes:
                 self.assignes[first.name] = []
+
+
+    # def fetch(
+    #     self,
+    #     frag: _Fragment,
+    #     buffer: Buffer,
+    #     upper_separators: str,
+    #     rxget: RxGet[Any] | None = None,  # type: ignore
+    #     rxput: RxPut[Any] | None = None,  # type: ignore
+    # ):
+    #     if frag.separators is not None:
+    #         if frag.hybrid_separators:
+    #             separators = frag.separators + upper_separators
+    #         else:
+    #             separators = frag.separators
+    #     else:
+    #         separators = upper_separators
+
+    #     token = buffer.next(separators)
+
+    #     if rxput is None:
+    #         def rxput(val):
+    #             self.assignes[frag.name] = val
+        
+    #     rxput(token.val)
+    #     token.apply()
 
     def fetch(
         self,
@@ -125,7 +153,7 @@ class Track:
         return Track(self.fragments.copy(), self.assignes.copy())
 
 
-@dataclass
+@dataclass(**safe_dcls_kw(slots=True))
 class Preset:
     tracks: dict[str, deque[_Fragment]] = field(default_factory=dict)
 
@@ -140,7 +168,7 @@ class Preset:
         return Mix(self)
 
 
-@dataclass
+@dataclass(**safe_dcls_kw(slots=True))
 class Mix:
     preset: Preset
     tracks: dict[str, Track] = field(init=False, default_factory=dict)
